Move model types to shared
