+++ client/src/services/feedbackServices.ts
@@ -1,9 +1,9 @@
 import { AxiosResponse, AxiosError } from 'axios';
 import { api } from 'client/utils/api';
-import { Feedback } from 'client/typings/feedback.typings';
 import { ServerError } from 'shared/typings/api/errors';
 import { FEEDBACK_ENDPOINT } from 'shared/constants/apiEndpoints';
 import { PostFeedbackResponse } from 'shared/typings/api/feedback';
+import { Feedback } from 'shared/typings/models/feedback';
 
 export const postFeedback = async (
   feedbackData: Feedback
