--- client/src/typings/feedback.typings.ts
@@ -1,4 +1,0 @@
-export interface Feedback {
-  feedback: string;
-  gameId: string;
-}
