+++ server/src/features/feedback/feedbackRepository.ts
@@ -1,6 +1,6 @@
 import { logger } from 'server/utils/logger';
 import { FeedbackModel } from 'server/features/feedback/feedbackSchema';
-import { Feedback } from 'server/typings/feedback.typings';
+import { Feedback } from 'shared/typings/models/feedback';
 
 export const saveFeedback = async (feedbackData: Feedback): Promise<void> => {
   const feedback = new FeedbackModel({
