+++ server/src/features/player-assignment/munkres/utils/buildSignupResults.ts
@@ -1,5 +1,5 @@
 import { Game } from 'shared/typings/models/game';
-import { Result } from 'server/typings/result.typings';
+import { Result } from 'shared/typings/models/result';
 import { SelectedGame, User } from 'shared/typings/models/user';
 
 export const buildSignupResults = (
