+++ server/src/features/player-assignment/utils/removeOverlapSignups.ts
@@ -1,8 +1,9 @@
 import moment from 'moment';
 import { logger } from 'server/utils/logger';
-import { UserSignup, Result } from 'server/typings/result.typings';
+import { UserSignup } from 'server/typings/result.typings';
 import { User } from 'shared/typings/models/user';
 import { findUsers, saveSignup } from 'server/features/user/userRepository';
+import { Result } from 'shared/typings/models/result';
 
 export const removeOverlapSignups = async (
   results: readonly Result[]
