+++ server/src/features/player-assignment/utils/saveResults.ts
@@ -1,6 +1,6 @@
 import { logger } from 'server/utils/logger';
 import { saveUserSignupResults } from 'server/features/player-assignment/utils/saveUserSignupResults';
-import { Result } from 'server/typings/result.typings';
+import { Result } from 'shared/typings/models/result';
 import { saveResult } from 'server/features/results/resultsRepository';
 
 export const saveResults = async (
