+++ server/src/features/results/test/resultsService.test.ts
@@ -1,7 +1,7 @@
 import mongoose from 'mongoose';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import { ResultsModel } from 'server/features/results/resultsSchema';
-import { Result } from 'server/typings/result.typings';
+import { Result } from 'shared/typings/models/result';
 import { saveResult } from 'server/features/results/resultsRepository';
 
 let mongoServer: MongoMemoryServer;
