+++ server/src/typings/feedback.typings.ts
@@ -1,11 +1,4 @@
 import mongoose from 'mongoose';
-import { Record, String, Static } from 'runtypes';
+import { Feedback } from 'shared/typings/models/feedback';
 
 export interface FeedbackDoc extends Feedback, mongoose.Document {}
-
-export const FeedbackRuntype = Record({
-  gameId: String,
-  feedback: String,
-});
-
-export type Feedback = Static<typeof FeedbackRuntype>;
