+++ server/src/typings/result.typings.ts
@@ -1,15 +1,11 @@
 import mongoose from 'mongoose';
 import { Game } from 'shared/typings/models/game';
 import { SelectedGame, User } from 'shared/typings/models/user';
+import { Result } from 'shared/typings/models/result';
 
 export interface ResultDoc extends Result, mongoose.Document {
   algorithm: string;
   message: string;
-}
-
-export interface Result {
-  username: string;
-  enteredGame: SelectedGame;
 }
 
 export interface UserSignup {
