+++ shared/typings/api/results.ts
@@ -1,4 +1,4 @@
-import { Result } from 'client/typings/result.typings';
+import { Result } from 'shared/typings/models/result';
 
 export interface GetResultsResponse {
   message: string;
