+++ shared/typings/models/feedback.ts
@@ -1,0 +1,8 @@
+import { Record, String, Static } from 'runtypes';
+
+export const FeedbackRuntype = Record({
+  gameId: String,
+  feedback: String,
+});
+
+export type Feedback = Static<typeof FeedbackRuntype>;
