+++ shared/typings/models/result.ts
