+++ shared/typings/models/settings.ts
